Moved all the GithubLogic shared between App and WebApp to a shared class. Introduced pattern, on how to share logic between App and WebApp. Possible with the final target, to merge them together.
