+++ application/src/main/java/dev/ikm/komet/app/WebApp.java
@@ -160,10 +160,9 @@ import static dev.ikm.komet.preferences.JournalWindowSettings.*;
  * @see LoginFeatureFlag
  * @see KometPreferences
  */
-public class WebApp extends Application {
+public class WebApp extends Application implements AppInterface  {
 
     private static final Logger LOG = LoggerFactory.getLogger(WebApp.class);
-    private static final String CHANGESETS_DIR = "changeSets";
     public static final String ICON_LOCATION = "/icons/Komet.png";
     public static final SimpleObjectProperty<AppState> state = new SimpleObjectProperty<>(STARTING);
     public static final SimpleObjectProperty<User> userProperty = new SimpleObjectProperty<>();
@@ -182,6 +181,17 @@ public class WebApp extends Application {
     private Image appIcon;
     private LandingPageController landingPageController;
     private EvtBus kViewEventBus;
+
+    AppGithub appGithub;
+
+    @Override
+    public LandingPageController getLandingPageController() {
+        return landingPageController;
+    }
+    @Override
+    public GitHubPreferencesDao getGitHubPreferencesDao() {
+        return gitHubPreferencesDao;
+    }
 
     /**
      * An entry point to launch the newer UI panels.
@@ -388,6 +398,8 @@ public class WebApp extends Application {
 
     @Override
     public void start(Stage stage) {
+        appGithub = new AppGithub(this);
+
         try {
             WebApp.primaryStage = stage;
             Thread.currentThread().setUncaughtExceptionHandler((thread, exception) ->
@@ -476,7 +488,7 @@ public class WebApp extends Application {
     public void stop() {
         LOG.info("Stopping application\n\n###############\n\n");
 
-        disconnectFromGithub();
+        appGithub.disconnectFromGithub();
 
         if (IS_DESKTOP) {
             // close all journal windows
@@ -635,9 +647,9 @@ public class WebApp extends Application {
         MenuItem infoMenuItem = new MenuItem("Info");
         infoMenuItem.setOnAction(actionEvent -> infoAction());
         MenuItem pullMenuItem = new MenuItem("Pull");
-        pullMenuItem.setOnAction(actionEvent -> executeGitTask(PULL));
+        pullMenuItem.setOnAction(actionEvent -> appGithub.executeGitTask(PULL));
         MenuItem pushMenuItem = new MenuItem("Sync");
-        pushMenuItem.setOnAction(actionEvent -> executeGitTask(SYNC));
+        pushMenuItem.setOnAction(actionEvent -> appGithub.executeGitTask(SYNC));
 
         exchangeMenu.getItems().addAll(infoMenuItem, pullMenuItem, pushMenuItem);
         return exchangeMenu;
@@ -711,7 +723,7 @@ public class WebApp extends Application {
             landingPageController = landingPageLoader.getController();
             landingPageController.getWelcomeTitleLabel().setText("Welcome " + user.getName());
             landingPageController.setSelectedDatasetTitle(PrimitiveData.get().name());
-            landingPageController.getGithubStatusHyperlink().setOnAction(_ -> connectToGithub());
+            landingPageController.getGithubStatusHyperlink().setOnAction(_ -> appGithub.connectToGithub());
 
             stage.setTitle("Landing Page");
             stage.setMaximized(true);
@@ -1061,257 +1073,6 @@ public class WebApp extends Application {
         exportStage.show();
     }
 
-    /**
-     * Prompts the user for GitHub preferences and repository information.
-     * <p>
-     * This method displays a dialog where the user can enter their GitHub credentials
-     * and repository URL. It creates a CompletableFuture that will be resolved with
-     * {@code true} if the user successfully enters valid credentials and connects,
-     * or {@code false} if they cancel the operation.
-     *
-     * @return A CompletableFuture that completes with true if valid GitHub preferences
-     *         were successfully provided by the user, or false if the user canceled the operation
-     */
-    private CompletableFuture<Boolean> promptForGitHubPrefs() {
-        // Create a CompletableFuture that will be completed when the user makes a choice
-        CompletableFuture<Boolean> future = new CompletableFuture<>();
-
-        // Show dialog on JavaFX thread
-        runOnFxThread(() -> {
-            GlassPane glassPane = new GlassPane(landingPageController.getRoot());
-
-            final JFXNode<Pane, GitHubPreferencesController> githubPreferencesNode = FXMLMvvmLoader
-                    .make(GitHubPreferencesController.class.getResource("github-preferences.fxml"));
-            final Pane dialogPane = githubPreferencesNode.node();
-            final GitHubPreferencesController controller = githubPreferencesNode.controller();
-            Optional<GitHubPreferencesViewModel> githubPrefsViewModelOpt = githubPreferencesNode
-                    .getViewModel("gitHubPreferencesViewModel");
-
-            controller.getConnectButton().setOnAction(actionEvent -> {
-                controller.handleConnectButtonEvent(actionEvent);
-                githubPrefsViewModelOpt.ifPresent(githubPrefsViewModel -> {
-                    if (githubPrefsViewModel.validProperty().get()) {
-                        glassPane.removeContent(dialogPane);
-                        glassPane.hide();
-                        future.complete(true); // Complete with true on successful connection
-                    }
-                });
-            });
-
-            controller.getCancelButton().setOnAction(_ -> {
-                glassPane.removeContent(dialogPane);
-                glassPane.hide();
-                future.complete(false); // Complete with false on cancel
-            });
-
-            glassPane.addContent(dialogPane);
-            glassPane.show();
-        });
-
-        return future;
-    }
-
-    /**
-     * Sets up the UI to reflect a disconnected GitHub state.
-     * <p>
-     * This method updates the GitHub status hyperlink in the landing page to show that
-     * the application is disconnected from GitHub. When clicked, the hyperlink will
-     * attempt to connect to GitHub by calling the {@link #connectToGithub()} method.
-     * <p>
-     * The method runs on the JavaFX application thread to ensure thread safety when
-     * updating UI components.
-     */
-    private void gotoGitHubDisconnectedState() {
-        runOnFxThread(() -> {
-            if (landingPageController != null) {
-                Hyperlink githubStatusHyperlink = landingPageController.getGithubStatusHyperlink();
-                githubStatusHyperlink.setText("Disconnected, Select to connect");
-                githubStatusHyperlink.setOnAction(event -> connectToGithub());
-            }
-        });
-    }
-
-    /**
-     * Sets up the UI to reflect a connected GitHub state.
-     * <p>
-     * This method updates the GitHub status hyperlink in the landing page to show that
-     * the application is successfully connected to GitHub. When clicked, the hyperlink
-     * will disconnect from GitHub by calling the {@link #disconnectFromGithub()} method.
-     * <p>
-     * The method runs on the JavaFX application thread to ensure thread safety when
-     * updating UI components.
-     */
-    private void gotoGitHubConnectedState() {
-        runOnFxThread(() -> {
-            if (landingPageController != null) {
-                Hyperlink githubStatusHyperlink = landingPageController.getGithubStatusHyperlink();
-                githubStatusHyperlink.setText("Connected");
-                githubStatusHyperlink.setOnAction(event -> disconnectFromGithub());
-            }
-        });
-    }
-
-    /**
-     * Executes a Git task, ensuring preferences are valid first.
-     * <p>
-     * This method performs the following operations:
-     * <ol>
-     *   <li>Verifies that the data store root is available</li>
-     *   <li>Creates a changeset folder if it doesn't exist</li>
-     *   <li>Validates GitHub preferences and prompts for them if missing</li>
-     *   <li>Creates and runs the appropriate GitTask based on the operation mode</li>
-     * </ol>
-     * If GitHub preferences are missing or invalid, this method will prompt the user
-     * to enter them before proceeding with the requested operation.
-     *
-     * @param mode The operation mode (CONNECT, PULL, or SYNC) that determines what
-     *             Git operations will be performed
-     */
-    private void executeGitTask(GitTask.OperationMode mode) {
-        Optional<File> optionalDataStoreRoot = ServiceProperties.get(ServiceKeys.DATA_STORE_ROOT);
-        if (optionalDataStoreRoot.isEmpty()) {
-            LOG.error("ServiceKeys.DATA_STORE_ROOT not provided.");
-            return;
-        }
-
-        final File changeSetFolder = new File(optionalDataStoreRoot.get(), CHANGESETS_DIR);
-        if (!changeSetFolder.exists()) {
-            if (!changeSetFolder.mkdirs()) {
-                LOG.error("Unable to create {} directory", CHANGESETS_DIR);
-                return;
-            }
-        }
-
-        // Check if GitHub preferences are valid first
-        if (!gitHubPreferencesDao.validate()) {
-            LOG.info("GitHub preferences missing or incomplete. Prompting user...");
-
-            // Prompt for preferences before proceeding
-            promptForGitHubPrefs().thenAccept(confirmed -> {
-                if (confirmed) {
-                    // Preferences entered successfully, now run the GitTask
-                    createAndRunGitTask(mode, changeSetFolder);
-                } else {
-                    LOG.info("User cancelled the GitHub preferences dialog");
-                }
-            });
-        } else {
-            // Preferences already valid, run the GitTask directly
-            createAndRunGitTask(mode, changeSetFolder);
-        }
-    }
-
-    /**
-     * Creates and runs a GitTask with the specified operation mode.
-     * <p>
-     * This helper method is called after GitHub preferences have been validated. It:
-     * <ol>
-     *   <li>Creates a new GitTask with the specified operation mode</li>
-     *   <li>Registers a success callback to update the UI state</li>
-     *   <li>Runs the task with progress tracking</li>
-     *   <li>Handles errors and updates the UI accordingly</li>
-     * </ol>
-     * The task is executed asynchronously through the ProgressHelper service to
-     * provide user feedback during long-running operations.
-     *
-     * @param operationMode The operation mode specifying which Git operations to perform
-     * @param changeSetFolder The folder where the Git repository is located
-     * @return A CompletableFuture that completes with true if the operation was successful,
-     *         or false if it failed or was cancelled
-     */
-    private CompletableFuture<Boolean> createAndRunGitTask(GitTask.OperationMode operationMode, File changeSetFolder) {
-        // Create a GitTask with only the connection success callback
-        GitTask gitTask = new GitTask(operationMode, changeSetFolder.toPath(), this::gotoGitHubConnectedState);
-
-        // Run the task
-        return ProgressHelper.progress(LANDING_PAGE_TOPIC, gitTask)
-                .whenComplete((result, throwable) -> {
-                    if (throwable != null) {
-                        LOG.error("Error during {} operation", operationMode, throwable);
-                        disconnectFromGithub();
-                    } else if (!result) {
-                        LOG.warn("{} operation did not complete successfully", operationMode);
-                        disconnectFromGithub();
-                    }
-                });
-    }
-
-    /**
-     * Initiates a connection to GitHub.
-     * <p>
-     * This method establishes a connection to GitHub by executing a GitTask in CONNECT mode.
-     * If successful, the UI will be updated to reflect the connected state, and the local
-     * Git repository will be initialized and configured with the remote origin.
-     * <p>
-     * If GitHub preferences are missing or invalid, the user will be prompted to
-     * enter them before the connection is established.
-     */
-    private void connectToGithub() {
-        LOG.info("Attempting to connect to GitHub...");
-        executeGitTask(CONNECT);
-    }
-
-    /**
-     * Disconnects from GitHub and cleans up local resources.
-     * <p>
-     * This method performs the following cleanup operations:
-     * <ol>
-     *   <li>Logs the disconnection attempt</li>
-     *   <li>Removes all GitHub-related preferences from user preferences</li>
-     *   <li>Deletes the local .git repository folder if it exists</li>
-     *   <li>Deletes the README.md file if it exists</li>
-     *   <li>Updates the UI to reflect the disconnected state</li>
-     * </ol>
-     * If any errors occur during this process, they are logged but do not prevent
-     * the disconnection from completing.
-     */
-    private void disconnectFromGithub() {
-        LOG.info("Disconnecting from GitHub...");
-
-        // Delete stored user preferences related to GitHub
-        try {
-            gitHubPreferencesDao.delete();
-            LOG.info("Successfully deleted GitHub preferences");
-        } catch (BackingStoreException e) {
-            LOG.error("Failed to delete GitHub preferences", e);
-        }
-
-        // Delete the .git folder and README.md if they exist
-        Optional<File> optionalDataStoreRoot = ServiceProperties.get(ServiceKeys.DATA_STORE_ROOT);
-        if (optionalDataStoreRoot.isPresent()) {
-            final File changeSetFolder = new File(optionalDataStoreRoot.get(), CHANGESETS_DIR);
-
-            // Delete .git folder
-            final File gitDir = new File(changeSetFolder, ".git");
-            if (gitDir.exists() && gitDir.isDirectory()) {
-                try {
-                    FileUtils.delete(gitDir, FileUtils.RECURSIVE);
-                    LOG.info("Successfully deleted .git folder at: {}", gitDir.getAbsolutePath());
-                } catch (IOException e) {
-                    LOG.error("Failed to delete .git folder at: {}", gitDir.getAbsolutePath(), e);
-                }
-            }
-
-            // Delete README.md file
-            final File readmeFile = new File(changeSetFolder, README_FILENAME);
-            if (readmeFile.exists() && readmeFile.isFile()) {
-                try {
-                    if (readmeFile.delete()) {
-                        LOG.info("Successfully deleted {} file at: {}", README_FILENAME, readmeFile.getAbsolutePath());
-                    } else {
-                        LOG.error("Failed to delete {} file at: {}", README_FILENAME, readmeFile.getAbsolutePath());
-                    }
-                } catch (SecurityException e) {
-                    LOG.error("Security exception while deleting {} file at: {}", readmeFile.getAbsolutePath(), e);
-                }
-            }
-        } else {
-            LOG.warn("Could not access data store root to delete .git folder and README.md");
-        }
-
-        // Update the UI state
-        gotoGitHubDisconnectedState();
-    }
 
     /**
      * Displays information about the current Git repository.
@@ -1336,17 +1097,17 @@ public class WebApp extends Application {
             return;
         }
 
-        final File changeSetFolder = new File(optionalDataStoreRoot.get(), CHANGESETS_DIR);
+        final File changeSetFolder = new File(optionalDataStoreRoot.get(), AppGithub.CHANGESETS_DIR);
         final File gitDir = new File(changeSetFolder, ".git");
 
         if (gitDir.exists()) {
             fetchAndShowRepositoryInfo(changeSetFolder);
         } else {
             // Prompt for preferences before proceeding
-            promptForGitHubPrefs().thenCompose(confirmed -> {
+            appGithub.promptForGitHubPrefs().thenCompose(confirmed -> {
                 if (confirmed) {
                     // Preferences entered successfully, now run the GitTask
-                    return createAndRunGitTask(CONNECT, changeSetFolder);
+                    return appGithub.createAndRunGitTask(CONNECT, changeSetFolder);
                 } else {
                     return CompletableFuture.completedFuture(false);
                 }
